# api_docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CampusConnect API",
    "description": "B2B SaaS platform for institutions and scholarships",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/institutions": {
      "get": {
        "tags": [
          "institutions"
        ],
        "summary": "Get Institutions",
        "description": "Get all institutions with optional state filter",
        "operationId": "get_institutions_api_v1_institutions_get",
        "parameters": [
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 2
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InstitutionResponse"
                  },
                  "title": "Response Get Institutions Api V1 Institutions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/institutions/{ipeds_id}": {
      "get": {
        "tags": [
          "institutions"
        ],
        "summary": "Get Institution",
        "description": "Get a single institution by IPEDS ID",
        "operationId": "get_institution_api_v1_institutions__ipeds_id__get",
        "parameters": [
          {
            "name": "ipeds_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Ipeds Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstitutionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/auth/register": {
      "post": {
        "tags": [
          "admin-auth"
        ],
        "summary": "Register",
        "description": "Register a new admin user",
        "operationId": "register_api_v1_admin_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminRegister"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/auth/login": {
      "post": {
        "tags": [
          "admin-auth"
        ],
        "summary": "Login",
        "description": "Login and get access token",
        "operationId": "login_api_v1_admin_auth_login_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_login_api_v1_admin_auth_login_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Token"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/auth/me": {
      "get": {
        "tags": [
          "admin-auth"
        ],
        "summary": "Get Me",
        "description": "Get current user info",
        "operationId": "get_me_api_v1_admin_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/profile/entity": {
      "get": {
        "tags": [
          "admin-profile"
        ],
        "summary": "Get My Entity",
        "description": "Get the institution or scholarship this admin manages",
        "operationId": "get_my_entity_api_v1_admin_profile_entity_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/profile/display-settings": {
      "get": {
        "tags": [
          "admin-profile"
        ],
        "summary": "Get Display Settings",
        "description": "Get display settings for my entity",
        "operationId": "get_display_settings_api_v1_admin_profile_display_settings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisplaySettingsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "admin-profile"
        ],
        "summary": "Update Display Settings",
        "description": "Update display settings for my entity",
        "operationId": "update_display_settings_api_v1_admin_profile_display_settings_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DisplaySettingsUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisplaySettingsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/images/upload": {
      "post": {
        "tags": [
          "admin-images"
        ],
        "summary": "Upload Image",
        "description": "Upload an image to DigitalOcean Spaces",
        "operationId": "upload_image_api_v1_admin_images_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_image_api_v1_admin_images_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/images/list": {
      "get": {
        "tags": [
          "admin-images"
        ],
        "summary": "List Images",
        "description": "List all images for this entity from DigitalOcean Spaces",
        "operationId": "list_images_api_v1_admin_images_list_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/images/{filename}": {
      "delete": {
        "tags": [
          "admin-images"
        ],
        "summary": "Delete Image",
        "description": "Delete an image from DigitalOcean Spaces",
        "operationId": "delete_image_api_v1_admin_images__filename__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "filename",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Filename"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AdminRegister": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          },
          "entity_type": {
            "type": "string",
            "title": "Entity Type"
          },
          "entity_id": {
            "type": "integer",
            "title": "Entity Id"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password",
          "entity_type",
          "entity_id"
        ],
        "title": "AdminRegister"
      },
      "AdminResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "entity_type": {
            "type": "string",
            "title": "Entity Type"
          },
          "entity_id": {
            "type": "integer",
            "title": "Entity Id"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "entity_type",
          "entity_id",
          "role",
          "is_active"
        ],
        "title": "AdminResponse"
      },
      "Body_login_api_v1_admin_auth_login_post": {
        "properties": {
          "grant_type": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "password"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grant Type"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "password": {
            "type": "string",
            "title": "Password"
          },
          "scope": {
            "type": "string",
            "title": "Scope",
            "default": ""
          },
          "client_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Id"
          },
          "client_secret": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Secret"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "Body_login_api_v1_admin_auth_login_post"
      },
      "Body_upload_image_api_v1_admin_images_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_image_api_v1_admin_images_upload_post"
      },
      "DisplaySettingsResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "entity_type": {
            "type": "string",
            "title": "Entity Type"
          },
          "entity_id": {
            "type": "integer",
            "title": "Entity Id"
          },
          "show_stats": {
            "type": "boolean",
            "title": "Show Stats"
          },
          "show_financial": {
            "type": "boolean",
            "title": "Show Financial"
          },
          "show_requirements": {
            "type": "boolean",
            "title": "Show Requirements"
          },
          "show_image_gallery": {
            "type": "boolean",
            "title": "Show Image Gallery"
          },
          "show_video": {
            "type": "boolean",
            "title": "Show Video"
          },
          "show_extended_info": {
            "type": "boolean",
            "title": "Show Extended Info"
          },
          "custom_tagline": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Custom Tagline"
          },
          "custom_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Custom Description"
          },
          "extended_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extended Description"
          },
          "layout_style": {
            "type": "string",
            "title": "Layout Style"
          },
          "primary_color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Color"
          }
        },
        "type": "object",
        "required": [
          "id",
          "entity_type",
          "entity_id",
          "show_stats",
          "show_financial",
          "show_requirements",
          "show_image_gallery",
          "show_video",
          "show_extended_info",
          "custom_tagline",
          "custom_description",
          "extended_description",
          "layout_style",
          "primary_color"
        ],
        "title": "DisplaySettingsResponse"
      },
      "DisplaySettingsUpdate": {
        "properties": {
          "show_stats": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Show Stats"
          },
          "show_financial": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Show Financial"
          },
          "show_requirements": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Show Requirements"
          },
          "show_image_gallery": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Show Image Gallery"
          },
          "show_video": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Show Video"
          },
          "show_extended_info": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Show Extended Info"
          },
          "custom_tagline": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Custom Tagline"
          },
          "custom_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Custom Description"
          },
          "extended_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extended Description"
          },
          "layout_style": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Layout Style"
          },
          "primary_color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Color"
          }
        },
        "type": "object",
        "title": "DisplaySettingsUpdate"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InstitutionResponse": {
        "properties": {
          "ipeds_id": {
            "type": "integer",
            "title": "Ipeds Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "city": {
            "type": "string",
            "title": "City"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "control_type": {
            "type": "string",
            "title": "Control Type"
          },
          "primary_image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Image Url"
          },
          "student_faculty_ratio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Student Faculty Ratio"
          },
          "size_category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size Category"
          },
          "locale": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Locale"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "ipeds_id",
          "name",
          "city",
          "state",
          "control_type",
          "id",
          "created_at",
          "updated_at"
        ],
        "title": "InstitutionResponse"
      },
...
      "Token": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "token_type"
        ],
        "title": "Token"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/admin/auth/login"
          }
        }
      }
    }
  }
}